Centralize compiler dependency plan lowering
Amp-Thread-ID: https://ampcode.com/threads/T-01a06089-1d42-711b-a1d1-cad8c8bff1f2
